feat(config): enhance environment configuration files with detailed comments and restructure integration modules
- Add comprehensive Chinese comments and explanations for `.env.example` and `.env.test.example`
- Rename and restructure Lark integration repositories and services for clarity and consistency
- Introduce typed and namespaced configuration management for Aliyun and Lark integrations
- Update Aliyun SMS service to use strongly typed config injection instead of environment variables
- Add config validation error messages for missing critical environment variables in Lark repositories
- Add unit tests for Lark config mapping from environment variables
- Refactor import paths and barrel files for common enums and integration modules
- Add `@nestjs/schedule` dependency to package manifest for scheduling functionalities
- Update documentation to reflect configuration key renaming and usage conventions
