+++ .env.example
@@ -9,10 +9,34 @@
  # Copyright (c) 2025 by ${git_name_email}, All Rights Reserved. 
 ### 
 
-# Database
+# =============================================================================
+# 数据库配置
+# =============================================================================
+# PostgreSQL 数据库连接字符串
+# 格式: postgresql://username:password@host:port/database_name?schema=public
+# 参数说明:
+#   - username: 数据库用户名
+#   - password: 数据库密码
+#   - host: 数据库服务器地址 (本地用 localhost，远程用 IP 或域名)
+#   - port: 数据库端口 (默认 5432)
+#   - database_name: 数据库名称
+#   - schema: 数据库模式 (默认 public)
+# 安全提示: 在生产环境中，请使用强密码并限制数据库访问权限
 DATABASE_URL="postgresql://username:password@localhost:5432/lulab_backend?schema=public"
 
-# Tencent Meeting Configuration
+# =============================================================================
+# 腾讯会议 API 配置
+# =============================================================================
+# 用于集成腾讯会议服务，实现会议创建、管理、录制等功能
+# 配置说明:
+#   - TENCENT_MEETING_APP_ID: 腾讯会议应用 ID (在腾讯会议开放平台申请)
+#   - TENCENT_MEETING_SDK_ID: SDK ID (用于客户端集成)
+#   - TENCENT_MEETING_SECRET_ID: 密钥 ID (用于 API 身份验证)
+#   - TENCENT_MEETING_SECRET_KEY: 密钥内容 (用于 API 签名加密)
+#   - TENCENT_MEETING_TOKEN: 用于接收和验证腾讯会议webhook的访问令牌
+#   - TENCENT_MEETING_ENCODING_AES_KEY: 消息加密密钥 (用于 Webhook 消息解密)
+# 获取方式: 登录腾讯会议开放平台 (https://meeting.tencent.com/open-api)
+# 安全提示: 密钥信息请妥善保管，不要在客户端代码中暴露
 TENCENT_MEETING_APP_ID="your_app_id"
 TENCENT_MEETING_SDK_ID="your_sdk_id"
 TENCENT_MEETING_SECRET_ID="your_secret_id"
@@ -20,12 +44,31 @@ TENCENT_MEETING_SECRET_KEY="your_secret_key"
 TENCENT_MEETING_TOKEN="your_webhook_token"
 TENCENT_MEETING_ENCODING_AES_KEY="your_encoding_aes_key"
 
-# Server Configuration
+# =============================================================================
+# 服务器配置
+# =============================================================================
+# 应用服务器端口和运行环境配置
+# 配置说明:
+#   - PORT: 服务器监听端口 (默认 3000)
+#   - NODE_ENV: 运行环境 (development 开发环境, production 生产环境, test 测试环境)
 PORT=3000
 NODE_ENV=development
 
 
-# SMTP邮件服务配置
+# =============================================================================
+# SMTP 邮件服务配置
+# =============================================================================
+# 用于发送系统邮件通知，如注册验证、密码重置、会议提醒等
+# 配置说明:
+#   - SMTP_HOST: SMTP 服务器地址
+#   - SMTP_PORT: SMTP 端口 (465 为 SSL 加密端口，587 为 TLS 端口)
+#   - SMTP_SECURE: 是否使用安全连接 (true 使用 SSL/TLS 加密)
+#   - SMTP_USER: 邮箱账号用户名
+#   - SMTP_PASS: 邮箱账号授权码 (注意：不是登录密码，需要在邮箱设置中生成)
+#   - SMTP_FROM: 发件人显示地址
+# 安全提示: 建议使用专用的邮件发送账号，并定期更新授权码
+
+# Gmail 配置示例
 SMTP_HOST=smtp.gmail.com
 SMTP_PORT=587
 SMTP_SECURE=false
@@ -64,19 +107,124 @@ ALIYUN_SMS_TEMPLATE_REGISTER=SMS_271525576
 ALIYUN_SMS_TEMPLATE_LOGIN=SMS_271525576
 ALIYUN_SMS_TEMPLATE_RESET=SMS_271525576
 
-# JWT配置
+# =============================================================================
+# JWT 身份验证配置
+# =============================================================================
+# 用于用户身份验证和会话管理的 JSON Web Token 设置
+# 配置说明:
+#   - JWT_SECRET: 访问令牌密钥，用于签名和验证 JWT 令牌
+#   - JWT_REFRESH_SECRET: 刷新令牌密钥，用于签名和验证刷新令牌
+#   - JWT_EXPIRES_IN: 访问令牌有效期 (15m = 15分钟)
+#   - JWT_REFRESH_EXPIRES_IN: 刷新令牌有效期 (7d = 7天)
+#
+# 🔐 JWT_SECRET 生成方法 (选择其中一种):
+#
+# 方法1 - Node.js 加密随机生成:
+#   node -e "console.log(require('crypto').randomBytes(64).toString('hex'))"
+#
+# 方法2 - OpenSSL 命令行:
+#   openssl rand -hex 64
+#
+# 方法3 - Python 随机生成:
+#   python3 -c "import secrets; print(secrets.token_hex(64))"
+#
+# 方法4 - Linux/macOS 系统命令:
+#   head -c 64 /dev/urandom | base64 | tr -d "=+/" | cut -c1-64
+#
+# 安全建议:
+#   - 密钥应使用强随机字符串，长度至少 32 字符
+#   - 生产环境请使用不同的访问密钥和刷新密钥
+#   - 定期轮换密钥以提高安全性
+#   - 访问令牌有效期不宜过长，刷新令牌可设置较长有效期
+#   - ⚠️  警告: 请勿使用过于简单的密钥
 JWT_SECRET=your-super-secret-jwt-key-change-this-in-production
 JWT_REFRESH_SECRET=your-super-secret-refresh-key-change-this-in-production
 JWT_EXPIRES_IN=15m
 JWT_REFRESH_EXPIRES_IN=7d
 
-# 飞书集成配置
+# =============================================================================
+# 飞书开放平台配置
+# =============================================================================
+# 飞书应用基础配置，用于身份验证和 API 调用
+# 获取方式: 登录飞书开放平台 (https://open.feishu.cn/)
+# 配置路径: 开发者后台 -> 应用 -> 凭证与基础信息
+# 权限要求: 需要开通相应的 API 权限，如用户管理、多维表格等
 LARK_APP_ID=your_lark_app_id
 LARK_APP_SECRET=your_lark_app_secret
-LARK_BITABLE_APP_TOKEN=your_bitable_app_token
-LARK_BITABLE_MEETING_TABLE_ID=your_meeting_table_id
-LARK_BITABLE_RECORDING_FILE_TABLE_ID=your_recording_file_table_id
-LARK_BITABLE_MEETING_USER_TABLE_ID=your_meeting_user_table_id
 
-# 数据库配置
-DATABASE_URL="postgresql://username:password@localhost:5432/lulab_db?schema=public"
+# =============================================================================
+# 飞书多维表格配置
+# =============================================================================
+# 用于会议数据存储和管理的飞书多维表格
+# 配置说明:
+#   - LARK_BITABLE_APP_TOKEN: 多维表格应用令牌 (在飞书多维表格中获取)
+#   - LARK_TABLE_*: 各个数据表的 ID，用于数据存储和查询
+# 数据表用途:
+#   - MEETING: 会议基础信息表
+#   - MEETING_USER: 会议用户信息表
+#   - MEETING_RECORD_FILE: 录制文件信息表
+#   - MEETING_MEMBER: 会议成员记录表
+#   - RECORDING_FILE: 录制文件信息表
+# 获取方式: 在飞书多维表格中创建相应表格后，从表格 URL 中提取 table ID
+LARK_BITABLE_APP_TOKEN=your_bitable_app_token
+LARK_TABLE_MEETING=your_meeting_table_id
+LARK_TABLE_MEETING_USER=your_meeting_user_table_id
+LARK_TABLE_MEETING_RECORD_FILE=your_recording_file_table_id
+LARK_TABLE_MEETING_MEMBER=your_meeting_member_table_id
+
+# =============================================================================
+# 数据库加密配置
+# =============================================================================
+# 用于敏感数据的数据库级加密，如用户个人信息、会议记录等
+# 要求: 
+#   - 密钥长度必须 32 字符以上
+#   - 建议使用随机生成的强密码，包含大小写字母、数字和特殊字符
+#   - 生产环境请使用专用的密钥管理服务
+# 安全提示: 
+#   - 此密钥用于数据库敏感字段加密，请妥善保管
+#   - 更改密钥会导致已加密数据无法解密，请谨慎操作
+#   - 建议定期轮换密钥并使用密钥版本管理
+ENCRYPTION_KEY=your-encryption-key-minimum-32-characters-long
+
+# =============================================================================
+# 阿里云短信服务配置
+# =============================================================================
+# 用于发送短信验证码，支持用户注册、登录、密码重置等场景
+# 
+# ⚠️  安全警告 ⚠️
+# 当前配置使用的是阿里云 AccessKey，存在安全风险
+# 强烈建议在生产环境中使用以下更安全的方案：
+#   1. RAM 角色 (推荐): https://help.aliyun.com/document_detail/378664.html
+#   2. STS 临时凭证: 动态获取临时访问凭证
+#   3. 阿里云凭证库: 自动管理凭证生命周期
+# 
+# 配置说明:
+#   - ALIBABA_CLOUD_ACCESS_KEY_ID: 阿里云访问密钥 ID
+#   - ALIBABA_CLOUD_ACCESS_KEY_SECRET: 阿里云访问密钥 Secret
+#   - ALIYUN_SMS_SIGN_NAME: 短信签名名称 (需在阿里云控制台申请并通过审核)
+#   - ALIYUN_SMS_TEMPLATE_*: 短信模板代码 (需在阿里云控制台申请并通过审核)
+# 
+# 短信模板要求:
+#   - 模板内容必须符合阿里云短信服务规范
+#   - 模板中的变量格式: ${code} 表示验证码
+#   - 不同场景使用不同模板，便于统计和管理
+# 
+# 获取方式: 登录阿里云控制台 -> 短信服务 -> 国内消息
+# 阿里云短信服务文档: https://help.aliyun.com/document_detail/55284.html
+
+# 注意：建议使用RAM角色或STS临时凭证，而不是直接配置AccessKey
+# 详见：https://help.aliyun.com/document_detail/378664.html
+# 阿里云凭证库使用的标准环境变量名称
+ALIBABA_CLOUD_ACCESS_KEY_ID=your-access-key-id
+ALIBABA_CLOUD_ACCESS_KEY_SECRET=your-access-key-secret
+
+# 阿里云短信签名（需要在阿里云控制台申请并审核通过）
+# 签名将显示在短信开头，如：【您的签名】您的验证码是 123456
+ALIYUN_SMS_SIGN_NAME=your-sms-sign-name
+
+# 阿里云短信模板代码（需要在阿里云控制台申请并审核通过）
+# 模板示例: 您的验证码是 ${code}，5分钟内有效，请勿泄露
+# 当前配置为同一个模板用于所有场景，建议为不同场景创建专用模板
+ALIYUN_SMS_TEMPLATE_REGISTER=SMS_271525576      # 用户注册场景
+ALIYUN_SMS_TEMPLATE_LOGIN=SMS_271525576         # 用户登录场景  
+ALIYUN_SMS_TEMPLATE_RESET=SMS_271525576           # 密码重置场景
