+++ .env.test.example
@@ -1,60 +1,134 @@
+# =============================================================================
 # 测试环境配置示例
-# 复制此文件为 .env.test 并填入实际值
+# =============================================================================
+# 复制此文件为 .env.test 并填入实际测试值
+# 注意：测试环境请使用与生产环境隔离的配置
 
+# =============================================================================
 # 数据库配置（分层测试数据库）
+# =============================================================================
+# PostgreSQL 数据库连接字符串（测试环境）
+# 建议为不同测试类型使用独立的数据库，避免数据干扰
 DATABASE_URL="postgresql://username:password@localhost:5432/lulab_test"
 DATABASE_URL_UNIT_TEST="postgresql://username:password@localhost:5432/lulab_unit_test"
 DATABASE_URL_INTEGRATION_TEST="postgresql://username:password@localhost:5432/lulab_integration_test"
 DATABASE_URL_SYSTEM_TEST="postgresql://username:password@localhost:5432/lulab_system_test"
 DATABASE_URL_E2E_TEST="postgresql://username:password@localhost:5432/lulab_e2e_test"
 
-# Redis配置（如果使用）
+# =============================================================================
+# Redis 配置
+# =============================================================================
+# 测试环境 Redis 连接（推荐使用 Docker 启动测试 Redis 实例）
+# 命令: docker run -d -p 6379:6379 --name test-redis redis:latest
 REDIS_URL="redis://localhost:6379"
 
-# JWT配置
-JWT_SECRET="your-test-jwt-secret"
-JWT_EXPIRATION="1d"
+# =============================================================================
+# JWT 身份验证配置
+# =============================================================================
+# 测试环境 JWT 配置（使用简单密钥便于测试）
+JWT_SECRET="test-jwt-secret-key-please-change-in-production"
+JWT_REFRESH_SECRET="test-jwt-refresh-secret-key-please-change"
+JWT_EXPIRES_IN="15m"
+JWT_REFRESH_EXPIRES_IN="7d"
 
-# 腾讯会议API配置
+# =============================================================================
+# 数据库加密配置
+# =============================================================================
+# 测试环境数据库加密密钥（32字符以上）
+ENCRYPTION_KEY="test-encryption-key-32-chars-min-length-required"
+
+# =============================================================================
+# 腾讯会议 API 配置（测试环境）
+# =============================================================================
+# 测试环境请使用腾讯会议沙箱环境或测试应用
+TENCENT_MEETING_APP_ID="test-app-id"
 TENCENT_MEETING_SDK_ID="test-sdk-id"
 TENCENT_MEETING_SECRET_ID="test-secret-id"
 TENCENT_MEETING_SECRET_KEY="test-secret-key"
-TENCENT_MEETING_APP_ID="test-app-id"
-TENCENT_MEETING_USER_ID="test-user-id"
+TENCENT_MEETING_TOKEN="test-webhook-token"
+TENCENT_MEETING_ENCODING_AES_KEY="test-encoding-aes-key-43-chars"
 
-# 飞书会议配置
+# =============================================================================
+# 飞书开放平台配置（测试环境）
+# =============================================================================
+# 测试环境请使用飞书测试应用
+LARK_APP_ID="test-lark-app-id"
+LARK_APP_SECRET="test-lark-app-secret"
+
+# =============================================================================
+# 飞书多维表格配置（测试环境）
+# =============================================================================
+# 测试环境多维表格配置（建议创建专用的测试表格）
+LARK_BITABLE_APP_TOKEN="test-bitable-app-token"
+LARK_TABLE_MEETING="test-meeting-table-id"
+LARK_TABLE_MEETING_USER="test-meeting-user-table-id"
+LARK_TABLE_MEETING_RECORD_FILE="test-meeting-record-file-table-id"
+LARK_TABLE_MEETING_MEMBER="test-meeting-member-table-id"
+
+# =============================================================================
+# SMTP 邮件服务配置（测试环境）
+# =============================================================================
+# 测试环境建议使用邮件测试服务，如 MailHog、MailCatcher 等
+# MailHog 启动命令: docker run -d -p 1025:1025 -p 8025:8025 --name test-mailhog mailhog/mailhog
+SMTP_HOST="localhost"
+SMTP_PORT="1025"
+SMTP_SECURE="false"
+SMTP_USER="test@lulabs.test"
+SMTP_PASS="test-password"
+SMTP_FROM="test@lulabs.test"
+
+# =============================================================================
+# 阿里云短信服务配置（测试环境）
+# =============================================================================
+# 测试环境建议使用阿里云短信测试签名和模板
+# 注意：测试环境每日有发送数量限制
+ALIBABA_CLOUD_ACCESS_KEY_ID="test-access-key-id"
+ALIBABA_CLOUD_ACCESS_KEY_SECRET="test-access-key-secret"
+ALIYUN_SMS_SIGN_NAME="测试签名"
+ALIYUN_SMS_TEMPLATE_REGISTER="SMS_test_template_code"
+ALIYUN_SMS_TEMPLATE_LOGIN="SMS_test_template_code"
+ALIYUN_SMS_TEMPLATE_RESET="SMS_test_template_code"
+
+# =============================================================================
+# 飞书会议配置（兼容配置）
+# =============================================================================
+# 向后兼容的飞书会议配置（将逐步迁移到 LARK_* 配置）
 FEISHU_APP_ID="test-feishu-app-id"
 FEISHU_APP_SECRET="test-feishu-app-secret"
 FEISHU_WEBHOOK_TOKEN="test-feishu-webhook-token"
-LARK_APP_ID="test-lark-app-id"
-LARK_APP_SECRET="test-lark-app-secret"
-LARK_TEST_APP_TOKEN="test-bitable-app-token"
-LARK_TEST_TABLE_ID="test-bitable-table-id"
 
-# 邮件服务配置
-SMTP_HOST="smtp.example.com"
-SMTP_PORT=587
-SMTP_USER="test@example.com"
-SMTP_PASS="test-password"
-
-# 阿里云SMS配置
-ALIYUN_ACCESS_KEY_ID="test-access-key"
-ALIYUN_ACCESS_KEY_SECRET="test-access-secret"
-ALIYUN_SMS_REGION="cn-hangzhou"
-ALIYUN_SMS_SIGN_NAME="test-sign"
-
-# 测试配置
+# =============================================================================
+# 测试执行配置
+# =============================================================================
+# 测试超时时间配置（毫秒）
 TEST_TIMEOUT=30000
 INTEGRATION_TEST_TIMEOUT=60000
 SYSTEM_TEST_TIMEOUT=120000
 E2E_TEST_TIMEOUT=60000
 
-# 网络配置
-# 确保当前服务器IP已添加到各平台应用的白名单
+# =============================================================================
+# 调试和日志配置
+# =============================================================================
+# API 调试和日志配置（测试环境建议开启）
 DEBUG_TENCENT_API=true
 LOG_API_RESPONSES=true
+DEBUG_LARK_API=true
+DEBUG_ALIYUN_SMS=true
 
-# 测试用户ID
-USER_ID=test-user-id-for-tests
-TEST_RECORDING_FILE_ID=real-test-recording-file-id
+# =============================================================================
+# 测试数据配置
+# =============================================================================
+# 测试用户和数据标识
+USER_ID="test-user-id-for-tests"
+TEST_RECORDING_FILE_ID="test-recording-file-id"
 TEST_TIME_RANGE_HOURS=24
+TEST_MEETING_ROOM_ID="test-meeting-room-id"
+TEST_ORGANIZATION_ID="test-organization-id"
+
+# =============================================================================
+# 网络和安全配置
+# =============================================================================
+# 测试环境网络配置
+# 确保测试服务器IP已添加到各平台测试应用的白名单
+ALLOW_TEST_NETWORK_ACCESS=true
+SKIP_IP_WHITELIST_CHECK=true
