+++ libs/common/index.ts
@@ -1,3 +1,3 @@
 // Common library barrel. Export shared types, utils, errors here.
-// Currently empty placeholder to establish import pattern.
 export * from './utils';
+export * from './enums';
