+++ libs/integrations/aliyun/aliyun.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { AliyunSmsService } from './aliyun-sms.service';
+import { aliyunConfig } from './config/aliyun.config';
 
 @Module({
-  imports: [ConfigModule],
+  imports: [ConfigModule.forFeature(aliyunConfig)],
   providers: [AliyunSmsService],
   exports: [AliyunSmsService],
 })
