+++ libs/integrations/aliyun/config/aliyun.config.ts
@@ -1,0 +1,14 @@
+import { registerAs, ConfigType } from '@nestjs/config';
+
+export const aliyunConfig = registerAs('aliyun', () => ({
+  sms: {
+    signName: process.env.ALIYUN_SMS_SIGN_NAME ?? '视算新里程科技',
+    templates: {
+      register: process.env.ALIYUN_SMS_TEMPLATE_REGISTER ?? 'SMS_271525576',
+      login: process.env.ALIYUN_SMS_TEMPLATE_LOGIN ?? 'SMS_271525576',
+      resetPassword: process.env.ALIYUN_SMS_TEMPLATE_RESET ?? 'SMS_271525576',
+    },
+  },
+}));
+
+export type AliyunConfig = ConfigType<typeof aliyunConfig>;
