+++ libs/integrations/aliyun/index.ts
@@ -1,2 +1,3 @@
 export * from './aliyun.module';
 export * from './aliyun-sms.service';
+export * from './config/aliyun.config';
