+++ libs/integrations/lark/index.ts
@@ -1,6 +1,7 @@
 export * from './lark.module';
 export * from './lark.client';
-export * from './bitable.service';
+export * from './services/bitable.service';
 export * from './repositories';
-export * from './lark.types';
+export * from './types';
 export * from './exceptions/lark.exceptions';
+export * from './config/lark.config';
