+++ libs/integrations/lark/repositories/meeting.repository.ts
@@ -1,23 +1,13 @@
-import { Injectable, Logger } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
-import { BitableService } from '../bitable.service';
+import { Inject, Injectable, Logger } from '@nestjs/common';
+import { ConfigType } from '@nestjs/config';
+import { BitableService } from '../services/bitable.service';
 import {
   CreateRecordResponse,
   UpdateRecordResponse,
   BitableField,
-} from '../lark.types';
-
-interface MeetingData {
-  platform: string;
-  subject: string;
-  meeting_id: string;
-  sub_meeting_id?: string;
-  meeting_code?: string;
-  start_time?: number;
-  end_time?: number;
-  operator?: string[];
-  creator?: string[];
-}
+} from '../types/lark.types';
+import { larkConfig } from '../config/lark.config';
+import { MeetingData } from '../types';
 
 /**
  * Repository for meeting-related Bitable operations
@@ -31,16 +21,14 @@ export class MeetingBitableRepository {
 
   constructor(
     private readonly bitableService: BitableService,
-    private readonly configService: ConfigService,
+    @Inject(larkConfig.KEY) private readonly cfg: ConfigType<typeof larkConfig>,
   ) {
-    this.appToken = this.configService.get<string>('LARK_BITABLE_APP_TOKEN')!;
-    this.tableId = this.configService.get<string>(
-      'LARK_BITABLE_MEETING_TABLE_ID',
-    )!;
+    this.appToken = this.cfg.bitable.appToken;
+    this.tableId = this.cfg.bitable.tableIds.meeting;
 
     if (!this.appToken || !this.tableId) {
       throw new Error(
-        'LARK_BITABLE_APP_TOKEN and LARK_BITABLE_MEETING_TABLE_ID must be configured in environment variables',
+        'LARK_BITABLE_APP_TOKEN and LARK_TABLE_MEETING must be configured in environment variables',
       );
     }
   }
