+++ libs/integrations/lark/types/meeting.types.ts
@@ -1,0 +1,11 @@
+export interface MeetingData {
+  platform: string;
+  subject: string;
+  meeting_id: string;
+  sub_meeting_id?: string;
+  meeting_code?: string;
+  start_time?: number;
+  end_time?: number;
+  operator?: string[];
+  creator?: string[];
+}
