--- scripts/test-lark-integration.ts
@@ -1,106 +1,0 @@
-#!/usr/bin/env node
-
-/**
- * 飞书多维表格集成测试脚本
- * 
- * 此脚本用于测试腾讯会议事件与飞书多维表格的集成
- * 运行前请确保已配置正确的环境变量：
- * - LARK_APP_ID
- * - LARK_APP_SECRET
- * - LARK_BITABLE_APP_TOKEN
- * - LARK_BITABLE_MEETING_TABLE_ID
- */
-
-import { NestFactory } from '@nestjs/core';
-import { ConfigService } from '@nestjs/config';
-import { TencentEventHandlerService } from '../src/tencent-meeting/services/tencent-event-handler.service';
-import { TencentMeetingModule } from '../src/tencent-meeting/tencent-meeting.module';
-import { TencentMeetingEvent } from '../src/tencent-meeting/types/tencent-webhook-events.types';
-
-async function testLarkIntegration() {
-  console.log('🚀 开始测试飞书多维表格集成...\n');
-
-  try {
-    // 创建应用实例
-    const app = await NestFactory.createApplicationContext(TencentMeetingModule);
-    
-    // 获取服务
-    const configService = app.get(ConfigService);
-    const eventHandler = app.get(TencentEventHandlerService);
-
-    // 检查配置
-    const appId = configService.get('LARK_APP_ID');
-    const appSecret = configService.get('LARK_APP_SECRET');
-    const bitableAppToken = configService.get('LARK_BITABLE_APP_TOKEN');
-    const bitableTableId = configService.get('LARK_BITABLE_MEETING_TABLE_ID');
-
-    console.log('📋 当前配置检查:');
-    console.log(`   LARK_APP_ID: ${appId ? '✅ 已配置' : '❌ 未配置'}`);
-    console.log(`   LARK_APP_SECRET: ${appSecret ? '✅ 已配置' : '❌ 未配置'}`);
-    console.log(`   LARK_BITABLE_APP_TOKEN: ${bitableAppToken ? '✅ 已配置' : '❌ 未配置'}`);
-    console.log(`   LARK_BITABLE_MEETING_TABLE_ID: ${bitableTableId ? '✅ 已配置' : '❌ 未配置'}\n`);
-
-    if (!appId || !appSecret || !bitableAppToken || !bitableTableId) {
-      console.log('⚠️  缺少必要的飞书配置，跳过实际API调用测试。');
-      console.log('💡 请参考 docs/LARK_INTEGRATION.md 进行配置。\n');
-      
-      // 测试事件处理逻辑（不调用实际API）
-      console.log('🔧 测试事件处理逻辑...');
-      
-      const mockEvent: TencentMeetingEvent = {
-        event: 'meeting.started',
-        trace_id: 'test_trace_' + Date.now(),
-        payload: [
-          {
-            operate_time: Date.now(),
-            operator: {
-              userid: 'test_user',
-              uuid: 'test_uuid',
-              user_name: '测试用户',
-              instance_id: 'test_instance',
-            },
-            meeting_info: {
-              meeting_id: 'test_meeting_' + Date.now(),
-              meeting_code: '123456789',
-              subject: '测试飞书集成会议',
-              creator: {
-                userid: 'creator_user',
-                uuid: 'creator_uuid',
-                user_name: '会议创建者',
-              },
-              meeting_type: 0,
-              start_time: Math.floor(Date.now() / 1000),
-              end_time: Math.floor(Date.now() / 1000) + 3600, // 1小时后
-            },
-          },
-        ],
-      };
-
-      try {
-        await eventHandler.handleEvent(mockEvent);
-        console.log('✅ 事件处理逻辑测试完成');
-      } catch (error) {
-        console.log('⚠️  事件处理逻辑测试完成（配置缺失时跳过）');
-      }
-
-    } else {
-      console.log('🎯 配置完整，可以测试实际API调用...\n');
-      
-      // 这里可以添加实际的API测试
-      console.log('✅ 飞书集成配置验证完成！');
-    }
-
-    await app.close();
-    
-  } catch (error) {
-    console.error('❌ 测试失败:', error.message);
-    process.exit(1);
-  }
-}
-
-// 如果直接运行此脚本
-if (require.main === module) {
-  testLarkIntegration().catch(console.error);
-}
-
-export { testLarkIntegration };
