+++ src/meeting/types/meeting.types.ts
@@ -53,6 +53,9 @@ export interface CreateMeetingFileParams {
   processingStatus: ProcessingStatus;
 }
 
+// Repository-layer alias for creating meeting file
+export type CreateMeetingFileData = CreateMeetingFileParams;
+
 /**
  * 会议记录查询参数
  */
@@ -62,6 +65,41 @@ export interface GetMeetingRecordsParams {
   endDate?: Date;
   page?: number;
   limit?: number;
+}
+
+/**
+ * 仓储层：会议记录创建数据（与数据库字段一致）
+ */
+export interface CreateMeetingRecordData {
+  platform: MeetingPlatform;
+  platformMeetingId: string;
+  title: string;
+  meetingCode: string;
+  type: MeetingType;
+  hostUserId: string;
+  hostUserName: string;
+  startTime: Date;
+  endTime: Date;
+  durationSeconds: number;
+  hasRecording: boolean;
+  recordingStatus: ProcessingStatus;
+  processingStatus: ProcessingStatus;
+  metadata?: any;
+}
+
+/**
+ * 仓储层：会议记录更新数据（与服务层参数等价）
+ */
+export type UpdateMeetingRecordData = UpdateMeetingRecordParams;
+
+/**
+ * 仓储层：会议文件更新数据
+ */
+export interface UpdateMeetingFileData {
+  fileName?: string;
+  content?: string;
+  processingStatus?: ProcessingStatus;
+  metadata?: any;
 }
 
 /**
