+++ src/verification/verification.module.ts
@@ -2,7 +2,6 @@ import { Global, Module } from '@nestjs/common';
 import { VerificationService } from './verification.service';
 import { VerificationRepository } from './repositories/verification.repository';
 import { AliyunModule } from '@libs/integrations/aliyun/aliyun.module';
-import { AliyunSmsService } from '@libs/integrations/aliyun/aliyun-sms.service';
 import { EmailModule } from '@/email/email.module';
 import { PrismaService } from '@/prisma.service';
 import { VerificationController } from './verification.controller';
@@ -11,12 +10,7 @@ import { VerificationController } from './verification.controller';
 @Module({
   imports: [AliyunModule, EmailModule],
   controllers: [VerificationController],
-  providers: [
-    VerificationService,
-    VerificationRepository,
-    AliyunSmsService,
-    PrismaService,
-  ],
+  providers: [VerificationService, VerificationRepository, PrismaService],
   exports: [VerificationService],
 })
 export class VerificationModule {}
